+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support to smart images
 - New defmodule to manipulate the url
 - Method generate instead of HOF to create client
+- Add support to build url with trim
+- Add support to build url with horizontal align (halign)
+- Add support to build url with vertial align (valign)
+- Add support to build url with smart parameter
+- Add support to get meta data from image
+- More tests
 
 ## [0.1.0] - 2018-05-02
 ### Added